Move SubmitPinFormComponent under src/lib/components/feed directory
